Using cosine similarity directly and moved regiser_handler
