+++ pyproject.toml
@@ -18,7 +18,6 @@ openai = "^0.27.8"
 rich = "^13.4.1"
 redbaron = "^0.9.2"
 tenacity = "^8.2.2"
-scipy = "^1.10.1"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.3.2"
