--- src/file_handler/handler_registry.py
@@ -1,9 +1,0 @@
-handler_registry = {}
-
-
-def register_handler(ext):
-    def decorator(cls):
-        handler_registry[ext] = cls
-        return cls
-
-    return decorator
